+++ src/lib/pages/validator-details/components/validator-overview/index.tsx
@@ -63,7 +63,7 @@ export const ValidatorOverview = ({
       </Flex>
       {!isMobile && (
         <Flex backgroundColor="gray.900" p={6} rounded={8} w="100%">
-          <RecentBlocksSection hasTitle />
+          <RecentBlocksSection validatorAddress={validatorAddress} />
         </Flex>
       )}
       <VotingPowerChart